llm-service/requirements.txt
Django>=5.1,<6.0
grpcio>=1.66,<2.0
protobuf>=5.27,<7.0
